refactor: request-handler-service to proxy-service
